refactor: langgraph nodes
