+++ docker-compose.yml
@@ -31,6 +31,7 @@ services:
 
   langfuse-server:
     image: langfuse/langfuse:2
+    restart: always
     depends_on:
       db:
         condition: service_healthy
